+++ docs/conf.py
@@ -60,7 +60,7 @@ copyright = "2024, Einar Forselv"
 author = "Einar Forselv"
 
 # The short X.Y version
-version = "3.0.1"
+version = "3.0.2"
 # The full version, including alpha/beta/rc tags
 release = version